Catch theme inconsistencies
The error thrown when a theme doesn't quite abide by a contract doesn't really give any pointers to what is causing the issue(s).

Adding this shows you exactly where in the contract is wrong and it's simply a matter of figuring out which theme has the extra or missing tokens and correcting it.

Would be nice to be able to log the name of the theme here for more direct pointers, but unsure how to accomplish it.